Update premake.sh
Add a (temporal) fix on the apt keys to make apt-install work without errors while installing dependencies. This can lead to ERROR: cannot launch node of type [robot_state_publisher/robot_state_publisher]: robot_state_publisher when not fixed, which is just related to apt-install not accepting the key that changed on the ROS side.